==> src/0_tidy.Rmd <==
---
title: "Building Tidy Data"
author: "Ian Dinwoodie"
date: "`r Sys.Date()`"
output:
  pdf_document: default
  html_document: default
---

```{r setup, include=FALSE}
library(tidyverse)
library(ggpubr)
knitr::opts_chunk$set(echo=TRUE)
```

# Loading the Data

Load the raw data and verify its dimensions and structure.

```{r}
df <- read.csv('../data/dogs.csv', header=TRUE, skipNul=TRUE)
dim(df)
str(df)
```

We see that we have 1095 responses across 26 fields. The columns names are not
quite serviceable in their current state, so we rename them for ease of use.

```{r}
names <- c(
  'acq_12_wo_or_less',
  'at_least_1yo',
  'age_yrs',
  'sex',
  'neutered',
  'train_6mo_or_less',
  'train_age',
  'train_class_count',
  'train_technique',
  'restr_device',
  'aggression',
  'fear_anxiety',
  'jumping',
  'barking',
  'coprophagia',
  'compulsion',
  'soil_when',
  'soil_how',
  'soil_where',
  'rep_materials',
  'hyperactive',
  'destructive',
  'escape',
  'mounting',
  'take_again',
  'owner_id'
)
colnames(df) <- names
str(df)
```

# Specifying Data Types

## Continious

We don't want to interpret every column as characters (`chr`), let's start by
specifying the continuous variables.

```{r}
df$age_yrs <- as.integer(df$age_yrs)
summary(df$age_yrs)
```

## Discrete

### Factors

We convert a bulk of the columns to factors. Before the conversion, we need to
deal with responses that are comma separated lists.

```{r}
# We separate each training age period into its own column.
df <- df %>%
  mutate(train_1_3_mo = as.factor(ifelse(
    train_age == "" | train_age == "I don't know", NA, ifelse(
      grepl('3 months or younger', train_age), TRUE, FALSE)))) %>%
  mutate(train_4_mo = as.factor(ifelse(
    train_age == "" | train_age == "I don't know", NA, ifelse(
      grepl('4 months', train_age), TRUE, FALSE)))) %>%
  mutate(train_5_6_mo = as.factor(ifelse(
    train_age == "" | train_age == "I don't know", NA, ifelse(
      grepl('5-6 months', train_age), TRUE, FALSE))))

# We also record the training start age for each dog.
df <- df %>%
  mutate(train_start_age = ordered(ifelse(
    train_age == "" | train_age == "I don't know", NA, ifelse(
      train_1_3_mo == TRUE, '1-3 mo', ifelse(
        train_4_mo == TRUE, '4 mo', '5-6 mo'))),
    levels=c('1-3 mo', '4 mo', '5-6 mo')))

# Convert training technique to reward or punishment.
df$train_technique <- ifelse(
    grepl('Rewarding', df$train_technique), 'reward', df$train_technique)
df$train_technique <- ifelse(
    grepl('combination', df$train_technique), 'punish', df$train_technique)
df$train_technique <- ifelse(
    grepl('Tough love', df$train_technique), 'punish', df$train_technique)
df$train_technique <- ifelse(
    df$train_technique == 'reward' | df$train_technique == 'punish',
    df$train_technique, NA)

# Assign training class count to maximum selected option.
df$train_class_count <- ifelse(
    grepl('1-3', df$train_class_count), '1-3', df$train_class_count)
df$train_class_count <- ifelse(
    grepl('4-6', df$train_class_count), '4-6', df$train_class_count)
df$train_class_count <- ifelse(
    grepl('7-9', df$train_class_count), '7-9', df$train_class_count)
df$train_class_count <- ifelse(
    grepl('10+', df$train_class_count), '10+', df$train_class_count)
df$train_class_count <- ifelse(
    grepl('10+', df$train_class_count), '10+', df$train_class_count)
is.na(df$train_class_count) <- df$train_class_count == "I don't know"
df$train_class_count <- ordered(
    df$train_class_count, levels=c('1-3', '4-6', '7-9', '10+'))
```

Now we perform the conversion to factor data type.

```{r}
factors <- c(
  'acq_12_wo_or_less',
  'at_least_1yo',
  'sex',
  'neutered',
#  'train_age',
#  'train_class_count',
  'train_technique',
  'restr_device',
  'aggression',
  'fear_anxiety',
  'jumping',
  'barking',
  'coprophagia',
  'compulsion',
  'soil_when',
  'soil_how',
  'soil_where',
  'rep_materials',
  'hyperactive',
  'destructive',
  'escape',
  'mounting',
  'owner_id'
)

for (c in factors) {
  df[, c] <- as.factor(df[, c])
}

str(df[, factors])
```

### Boolean

It's clear that some factor columns can be converted to boolean (i.e., logical).

```{r}
df <- df %>%
  mutate(at_least_1yo = ifelse(at_least_1yo == 'Yes', TRUE, FALSE)) %>%
  mutate(neutered = ifelse(neutered == 'Yes', TRUE, FALSE)) %>%
  mutate(acq_12_wo_or_less = ifelse(
    acq_12_wo_or_less == "I don't know", NA, ifelse(
      acq_12_wo_or_less == 'Yes', TRUE, FALSE)))
```

# Deriving Columns

We derive some columns for ease of use and improved clarity, especially when
responses are comma separated lists.

```{r}
df <- df %>%
  mutate(male = ifelse(sex == 'Male', FALSE, TRUE)) %>%
  mutate(device_used = ifelse(
    restr_device == "", NA, ifelse(
      grepl('No devices were employed', restr_device), FALSE, TRUE)))

# Derive a column for each restraining device.
df$buckle_collar <- ifelse(
  is.na(df$device_used), NA, ifelse(
    grepl('Buckle collar', df$restr_device), TRUE, FALSE))
df$martingale <- ifelse(
  is.na(df$device_used), NA, ifelse(
    grepl('Martingale collar', df$restr_device), TRUE, FALSE))
df$slip_collar <- ifelse(
  is.na(df$device_used), NA, ifelse(
    grepl('Nylon slip collar', df$restr_device), TRUE, FALSE))
df$shock_collar <- ifelse(
  is.na(df$device_used), NA, ifelse(
    grepl('Electric shock collar', df$restr_device), TRUE, FALSE))
df$harness <- ifelse(
  is.na(df$device_used), NA, ifelse(
    grepl('Harness', df$restr_device), TRUE, FALSE))
df$harness <- ifelse(
  is.na(df$device_used), NA, ifelse(
    grepl('harness', df$restr_device), TRUE, df$harness))
df$head_halter <- ifelse(
  is.na(df$device_used), NA, ifelse(
    grepl('Head halter', df$restr_device), TRUE, FALSE))
df$choke_collar <- ifelse(
  is.na(df$device_used), NA, ifelse(
    grepl('Metal \"choke\" collar', df$restr_device), TRUE, FALSE))
df$prong_collar <- ifelse(
  is.na(df$device_used), NA, ifelse(
    grepl('Prong collar', df$restr_device), TRUE, FALSE))
df$no_devices <- ifelse(
  is.na(df$device_used), NA, ifelse(
    grepl('No devices were employed', df$restr_device), TRUE, FALSE))
```

# Response Complexity Reductions

To start, we reduce the behavior problems to boolean indicators.

```{r}
df <- df %>%
  mutate(aggression = ifelse(aggression == "", FALSE, TRUE)) %>%
  mutate(fear_anxiety = ifelse(fear_anxiety == "", FALSE, TRUE)) %>%
  mutate(jumping = ifelse(jumping == "", FALSE, TRUE)) %>%
  mutate(barking = ifelse(barking == "", FALSE, TRUE)) %>%
  mutate(coprophagia = ifelse(coprophagia == "", FALSE, TRUE)) %>%
  mutate(compulsion = ifelse(compulsion == "", FALSE, TRUE)) %>%
  mutate(house_soiling = ifelse(
    soil_when != "" | soil_how != "" | soil_where != "", FALSE, TRUE)) %>%
  mutate(rep_materials = ifelse(rep_materials == "", FALSE, TRUE)) %>%
  mutate(hyperactive = ifelse(hyperactive == "", FALSE, TRUE)) %>%
  mutate(destructive = ifelse(destructive == "", FALSE, TRUE)) %>%
  mutate(escape = ifelse(escape == "", FALSE, TRUE)) %>%
  mutate(mounting = ifelse(mounting == "", FALSE, TRUE))

str(df)
```

# Accounting for the Human-Factor

## Training Methods

It is possible that the training technique that participants have reported
(e.g., reward-based training) is at odds with the training devices that were
employed.

```{r}
devices <- c(
  'buckle_collar',
  'martingale',
  'slip_collar',
  'shock_collar',
  'harness',
  'head_halter',
  'choke_collar',
  'prong_collar'
)

plot_list <- list()
for (i in 1:length(devices)) {
  p <- ggplot(data=subset(df, !is.na(train_technique)),
              aes_string(x=devices[i],fill='train_technique')) +
    geom_bar() +
    theme(legend.position="none")
  plot_list[[i]] <- p
}
ggarrange(plotlist=plot_list, ncol=3, nrow=3, common.legend=TRUE,
          legend="bottom")
```

The plot of the martingale usage is a good example of the anticipated trend:
several participants using a punishment-based device have indicated reward-based
training. Let's calculate the number of dogs subjected to a punishing device
when their owner believed they were using reward based training and apply the
necessary adjustments.

```{r}
print('Training techniques:')
summary(df$train_technique)
cnt <- summary(df$train_technique)[2]

df <- df %>%
  mutate(adj_train_technique = factor(ifelse(
    is.na(train_technique), NA, ifelse(
      martingale | slip_collar | shock_collar | choke_collar | prong_collar,
      'punish', 'reward')))) %>%
  mutate(punish_device = factor(ifelse(
    martingale | slip_collar | shock_collar | choke_collar | prong_collar, TRUE,
    FALSE)))

print('Training techniques (adjusted):')
summary(df$adj_train_technique)

diff <- cnt - summary(df$adj_train_technique)[2]
print(paste('Delta:', diff))
```

Let's reconstruct the plot above to visually see the adjustment.

```{r}
plot_list <- list()
for (i in 1:length(devices)) {
  p <- ggplot(data=subset(df, !is.na(adj_train_technique)),
              aes_string(x=devices[i],fill='adj_train_technique')) +
    geom_bar() +
    theme(legend.position="none")
  plot_list[[i]] <- p
}
ggarrange(plotlist=plot_list, ncol=3, nrow=3, common.legend=TRUE,
          legend="bottom")
```
We can see that the employment of a device that punishes is reflected in the
training technique.

# Dropping Excess Data

## Applying Inclusion Criteria

```{r}
df <- filter(df, at_least_1yo == TRUE, age_yrs >= 1, age_yrs <= 35)
dim(df)
length(unique(df$owner_id))
```

We dropped 72 responses for dogs and 28 owners as a result of the inclusion
criteria.

## Dropping Columns

Drop columns that serve no purpose with the analysis.

```{r}
df <- subset(df, select=-c(
  take_again, # survey software logic variable
  soil_when,
  soil_how,
  soil_where,
  at_least_1yo, # survey software logic variable
  sex, # replaced with a male column
  restr_device, # devices moved into their own columns
  no_devices, # mirrors the device_used column
  train_age # replaced with extracted start age and age range columns
  ))
```

# Final Summary

Take a last look at the data before saving it to disk.

```{r}
dim(df)
summary(df)
```

# Saving the Tidy Data

Save the data to a file in RDS format so that the data types are saved and so
that the output is compressed.

```{r}
saveRDS(df, '../data/tidy.Rds')
```
