==> src/1_exploratory.Rmd <==
---
title: "Exploratory Data Analysis"
author: "Ian Dinwoodie"
date: "`r Sys.Date()`"
output:
  pdf_document: default
  html_document: default
---

```{r setup, include=FALSE}
library(tidyverse)
library(ggpubr)
library(GGally)
library(e1071) 
knitr::opts_chunk$set(echo=TRUE)
```

# Loading the Data

Load the raw data and verify its dimensions and structure.

```{r}
df <- readRDS('../data/tidy.Rds')
dim(df)
str(df)
```

# Basic Exploration

## Owner Identifier

The owner identifier is used to calculate the number of owners.

```{r}
# Number of unique owners after inclusion criteria.
length(unique(df$owner_id))
```

It is also used to calculate the number of dogs per household.

```{r}
summary(plyr::count(df, 'owner_id'))
```

We see the median number of dogs per household is 1 (range: 1 to 8). Now we can
drop the column to simply the data set.

```{r}
df <- subset(df, select=-c(owner_id))
```

## Overview of Data Set

Before we look at the data, let's add a basic behavior problem indicator column.

```{r}
df <- df %>%
  mutate(behav_problem = ifelse(
    aggression | fear_anxiety | jumping | barking | coprophagia | compulsion
    | house_soiling | rep_materials | hyperactive | destructive | escape
    | mounting, TRUE, FALSE))

summary(df$behav_problem)
```

Let's take a look at the data set as we head toward analysis.

```{r}
summary(df)
```

Notable observations:

* Median dog age is 7 yrs (range: 1 to 19 yrs).
  * More than half (54.4%) were acquired at 12 weeks or less.
* A majority (87.1%) of dogs were neutered.
* The gender split is nearly even with 48.6% males.
* About half of the dogs (48.3%) attended training at 6 months old or earlier
(i.e., puppy training).
  * About half (47.4%) of which started attending in the 1-3 month range.
  * A majority (87.4%) of the dogs that attended puppy training were subject to
  some form of restraining device.
    * The buckle collar was the most popular device at 47.6% usage.
    * The shock collar was the least popular at 1.8% usage.
  * A vast majority (89.1%) were believed to have been subjected to reward-based
  training.
    * Correcting for punishing restraint devices, only 64.0% were truly subject
    to reward based training; a 25.1% difference!
* A vast majority of dogs (99.3%) were reported to exhibit at least one type of
problematic behavior.
  * The top 3 most frequent behavior problems were house soiling, fear/anxiety,
  aggression.
  * The 3 least frequent behavior problems were hyperactivity, destruction, and
  mounting.

We create pairwise scatter plots for columns that all participants were
presented (i.e., no NA responses) and we exclude the individual behavior problem
columns for brevity.

```{r, fig.height=5, fig.width=5}
df %>%
  ggpairs(columns=c('acq_12_wo_or_less', 'age_yrs', 'neutered',
                    'train_6mo_or_less', 'male'),
          mapping=ggplot2::aes(color=behav_problem),
          diag=list(discrete='barDiag',
                    continuous=wrap('densityDiag', alpha=0.5)),
          legend=1,
          progress=FALSE) +
  theme(legend.position='bottom')
```

Our control group consists of the dogs that did not attend puppy training. We
can compare variable distributions across the experimental and control groups by
looking at the graphs along the `train_6mo_or_less` row. Thankfully, we see that
the distributions between the two groups for the plotted columns are roughly
equivalent.

## Continuous Variables

The age of the dog is the only continuous variable we are working with.

```{r}
ggplot(df, aes(age_yrs)) + geom_bar(fill='steelblue')
skewness(df$age_yrs)
```

We see a slight right skew in the plot. Let's try to center it by applying a log
transform.

```{r}
df <- df %>%
  mutate(log_age_yrs = log(age_yrs))
ggplot(df, aes(log_age_yrs)) + geom_bar(fill='steelblue')
skewness(df$log_age_yrs)

df <- subset(df, select=-c(log_age_yrs))
```

We see that the log transform resulted in a greater absolute skew, so we drop
the transformed column and rely on the original.

## Discrete Variables

### Independent Variables

```{r, fig.height=5, fig.width=5}
vars <- c(
  'acq_12_wo_or_less',
  'neutered',
  'train_6mo_or_less',
  'male',
  'train_1_3_mo',
  'train_4_mo',
  'train_5_6_mo',
  'train_start_age',
  'device_used',
  'buckle_collar',
  'martingale',
  'slip_collar',
  'shock_collar',
  'harness',
  'head_halter',
  'choke_collar',
  'prong_collar',
  'house_soiling',
  'adj_train_technique'
)

plot_list <- list()
for (i in 1:length(vars)) {
  col <- vars[i]
  p <- df %>%
    select(col) %>%
    drop_na(col) %>%
    ggplot(aes_string(x = col)) +
    geom_bar(fill='steelblue')
  plot_list[[i]] <- p
}
ggarrange(plotlist=plot_list, ncol=4, nrow=5)
```

### Dependent Variables

```{r, fig.height=4, fig.width=5}
outcomes <- c(
  'aggression',
  'fear_anxiety',
  'jumping',
  'barking',
  'coprophagia',
  'compulsion',
  'rep_materials',
  'hyperactive',
  'destructive',
  'escape',
  'mounting',
  'house_soiling'
)

plot_list <- list()
for (i in 1:length(outcomes)) {
  col <- outcomes[i]
  p <- df %>%
    select(col) %>%
    drop_na(col) %>%
    ggplot(aes_string(x = col)) +
    geom_bar(fill='steelblue')
  plot_list[[i]] <- p
}
ggarrange(plotlist=plot_list, ncol=4, nrow=3)
```

# Exploring Trends and Relationships

## Sex and Neuter Status

It's common to want to know the split of neuter status by sex, so let's generate
those numbers now.

```{r}
xtab <- xtabs(~male+neutered, data=df)
print(xtab)
```

## Control vs Experimental Group

Our control group consists of the dogs who did not attend puppy training and our
experimental group consists of those who did. Let's look at the variables common
to both groups with the plot color indicating the presence of a behavior
problem. Since we know a vast majority of dogs have at least one behavior
problem, we need to look for trends in individual behavior problems for the
plots to be useful.

```{r, fig.height=10, fig.width=5}
# Generate plots for each attribute split by a simple predictor.
pred <- 'train_6mo_or_less'
attribs <- c(
  'acq_12_wo_or_less',
  'age_yrs',
  'male',
  'neutered'
)
attribs <- sort(attribs)
outcomes <- sort(outcomes)

plot_list <- list()
cnt <- 1
labels <- NULL
for (i in 1:length(outcomes)) {
  outcome <- outcomes[i]
  for (j in 1:length(attribs)) {
    attrib <- attribs[j]
    p <- df %>%
      drop_na(attrib) %>%
      select(attrib, outcome, pred) %>%
      ggplot(aes_string(x=attrib, fill=pred)) +
      geom_bar(position = position_dodge(0.9)) +
      labs(fill=pred) +
      theme(legend.position='none') +
      facet_grid(as.formula(paste0('.~', outcome)))
    plot_list[[cnt]] <- p
    cnt <- cnt + 1
    labels <- c(labels, outcome)
  }
}
ggarrange(plotlist=plot_list, ncol=4, nrow=12, common.legend=TRUE,
          font.label=list(size=10), vjust=0.75, legend='bottom', labels=labels)
```

> Note: For each single plot the behavior problem is indicated by the label in
the top left corner. The left facet is the group of dogs without the behavior
problem and the right are the dogs with the behavior problem. Within each facet
the color indicates control (red) or experiment (blue) grouping.

## Within the Experimental Group

Within the experimental group we are curious to see the impact of various
training techniques and restrain devices on behavior problem occurrence. We
start by isolating the experimental group.

```{r}
df_exp <- df %>%
  filter(train_6mo_or_less == TRUE)
summary(df_exp)
```

Now we look at the impact of training age and frequency.

```{r, fig.height=10, fig.width=5}
attribs <- c(
  'train_1_3_mo',
  'train_4_mo',
  'train_5_6_mo',
  'train_start_age',
  'adj_train_technique'
)
attribs <- sort(attribs)

plot_list <- list()
cnt <- 1
labels <- NULL
for (i in 1:length(outcomes)) {
  outcome <- outcomes[i]
  for (j in 1:length(attribs)) {
    attrib <- attribs[j]
    p <- df %>%
      drop_na(attrib) %>%
      select(attrib, outcome) %>%
      ggplot(aes_string(x=attrib, fill=outcome)) +
      geom_bar(position = position_dodge(0.9)) +
      labs(fill='has this behavoior problem') +
      theme(legend.position='none')
    plot_list[[cnt]] <- p
    cnt <- cnt + 1
    labels <- c(labels, outcome)
  }
}
ggarrange(plotlist=plot_list, ncol=5, nrow=12, common.legend=TRUE,
          font.label=list(size=10), vjust=0.75, legend='bottom', labels=labels)
```
Next, we try to visual the impact (if any) of a non-punishing restraining device
choice. We also include the overall `device_used` column to see if there is a
trend observed for restraining devices as a whole.

```{r, fig.height=10, fig.width=5}
attribs <- c(
  'device_used',
  'buckle_collar',
  'harness',
  'head_halter'
)

plot_list <- list()
cnt <- 1
labels <- NULL
for (i in 1:length(outcomes)) {
  outcome <- outcomes[i]
  for (j in 1:length(attribs)) {
    attrib <- attribs[j]
    p <- df %>%
      drop_na(attrib) %>%
      select(attrib, outcome) %>%
      ggplot(aes_string(x=attrib, fill=outcome)) +
      geom_bar(position = position_dodge(0.9)) +
      labs(fill='has this behavoior problem') +
      theme(legend.position='none')
    plot_list[[cnt]] <- p
    cnt <- cnt + 1
    labels <- c(labels, outcome)
  }
}
ggarrange(plotlist=plot_list, ncol=4, nrow=12, common.legend=TRUE,
          font.label=list(size=10), vjust=0.75, legend='bottom', labels=labels)
```

Last, we look at the impact of punishing restraining devices.

```{r, fig.height=10, fig.width=5}
attribs <- c(
  'martingale',
  'slip_collar',
  'shock_collar',
  'choke_collar',
  'prong_collar'
)
attribs <- sort(attribs)

plot_list <- list()
cnt <- 1
labels <- NULL
for (i in 1:length(outcomes)) {
  outcome <- outcomes[i]
  for (j in 1:length(attribs)) {
    attrib <- attribs[j]
    p <- df %>%
      drop_na(attrib) %>%
      select(attrib, outcome) %>%
      ggplot(aes_string(x=attrib, fill=outcome)) +
      geom_bar(position = position_dodge(0.9)) +
      labs(fill='has this behavoior problem') +
      theme(legend.position='none')
    plot_list[[cnt]] <- p
    cnt <- cnt + 1
    labels <- c(labels, outcome)
  }
}
ggarrange(plotlist=plot_list, ncol=5, nrow=12, common.legend=TRUE,
          font.label=list(size=10), vjust=0.75, legend='bottom', labels=labels)
```
