==> src/3_alt_models.Rmd <==
---
title: "Investigating Alternate Models"
author: "Ian Dinwoodie"
date: "`r Sys.Date()`"
output:
  pdf_document: default
  html_document: default
---

```{r setup, include=FALSE}
library(tidyverse)
library(car)
knitr::opts_chunk$set(echo=TRUE)
```

# Workspace Setup

## Loading the Data

Load the raw data and verify its dimensions and structure.

```{r}
df <- readRDS('../data/tidy.Rds')
dim(df)
summary(df)
```

## Preparing Common Variables and Functions

```{r}
outcomes <- c(
  'aggression',
  'fear_anxiety',
  'jumping',
  'barking',
  'coprophagia',
  'compulsion',
  'house_soiling',
  'rep_materials',
  'hyperactive',
  'destructive',
  'escape',
  'mounting'
)
outcomes <- sort(outcomes)
```

```{r}
apply_min_xtab <- function(df, outcome, cutoff=10)
{
  drops <- NULL
  for (col in names(df)) {
    if (col == outcome) next
    if (is.integer(df[,col])) next
    
    xtab <- table(df[,col], df[,outcome])
    if (min(xtab) < cutoff) {
      drops <- c(drops, col)
      break
    }
  }
  
  if (length(drops) > 0) {
    cat('\nDropped from model due to insufficient responses:\n')
    cat(drops)
    cat('\n')
  }
  
  return(df[, !(names(df) %in% drops)])
}
```

```{r}
df_exp <- df %>%
  filter(train_6mo_or_less == TRUE)
summary(df_exp)
```

# Alternate Models

## Exploring Training Methods and Equipment

With regard to training methods, the following question was presented to
participants.

> At puppy training classes, what training techniques were used?
> - Rewarding techniques (e.g., treats, praise, pets)
> - Tough love techniques (e.g., yelling, bopping on the nose, swatting on the rump, alpha rolls (pinning on back until dog submits), use of aversive collars (e.g., shock, prong, choke), jerking on the leash, water spraying, scruffing)
> - A combination of rewarding and tough love techniques

However, after discussing with clinicians, the unanimous decision was that all
training methods that involved some form of punishment could be considered
punishment-based. This modification can be seen in the first notebook for this
study (`0_tidy.Rmd`) and the result can be seen by looking at the
`train_technique` column.

```{r}
summary(df_exp$train_technique)
```

In addition, we also presented the following questions to participants:

> What restraining/training devices were employed?
> - Nylon slip collar
> - Buckle collar
> - Head halter (with nose band)
> - Harness (around chest)
> - Metal "choke" collar
> - Prong collar
> - Martingale collar
> - Electric shock collar
> - No devices were employed

We exclude the "Other" response from the list above for simplicity since the
only actionable submissions were for harnesses. The harness count was updated
accordingly.

These devices generally fall into two caregories: punishing and non-punishing.

> Punishing:
> - Metal "choke" collar
> - Prong collar
> - Martingale collar
> - Electric shock collar
> - Nylon slip collar
> Non-punishing:
> - Buckle collar
> - Head halter (with nose band)
> - Harness (around chest)

We calculated the number of dogs exposed to these punishing devices in our
initial notebook.

```{r}
summary(df_exp$punish_device)
```

### Analysis of Grouped Devices

The `punish_device` variable is not enough for our "grouped" analysis since it
does not account for dogs being exposed to no devices. Therefore, we create a
device group (`device_group`) column with three values: punish, non-punish,
none.

```{r}
df_exp <- df_exp %>%
  mutate(device_group = as.factor(ifelse(
    device_used == FALSE, 'none', ifelse(
      punish_device == TRUE, 'punish', 'non_punish'))))

summary(df_exp$device_group)
```

```{r}
df_exp <- df_exp %>%
  mutate(reward = ifelse(
    is.na(train_technique), NA, ifelse(
      train_technique == 'reward', TRUE, FALSE)))

common_params <- c(
  'age_yrs',
  'male',
  'neutered',
  'acq_12_wo_or_less',
  'train_1_3_mo',
  'train_4_mo',
  'train_5_6_mo',
  'train_class_count'
)

glm_attribs <- c(
  common_params,
  'reward',
  'device_group'
)

print(glm_attribs)
```

```{r}
set.seed(1)
for (outcome in outcomes) {
  cat(paste(replicate(80, '-'), collapse=''))
  cat(paste0('\n', outcome, '\n'))
  f <- as.formula(paste0(outcome, '~', '.'))
  
  df_tmp <- df_exp[,c(outcome, glm_attribs)]
  df_tmp <- apply_min_xtab(df_tmp, outcome)
  
  glm_fit <- glm(f, data=df_tmp, family='binomial')
  print(summary(glm_fit))
  print(exp(cbind(OR=coef(glm_fit), suppressMessages(confint(glm_fit)))))
  cat('\nVIF:\n')
  print(car::vif(glm_fit))
  cat('\n')
}
```

### Considering only Punishing Devices

However, since harnesses, buckle collars, and head halters are neither
punishment nor reward is it worth looking at this "non-punishment" device group?
We don't expect any of these devices to "help" and that is confirmed above. By
including them we are just diluting our models. Instead, we just want to know
if punishing devices were used and what their impacts were.

```{r}
glm_attribs <- c(
  common_params,
  'reward',
  'punish_device'
)

set.seed(1)
for (outcome in outcomes) {
  cat(paste(replicate(80, '-'), collapse=''))
  cat(paste0('\n', outcome, '\n'))
  f <- as.formula(paste0(outcome, '~', '.'))
  
  df_tmp <- df_exp[,c(outcome, glm_attribs)]
  df_tmp <- apply_min_xtab(df_tmp, outcome)
  
  glm_fit <- glm(f, data=df_tmp, family='binomial')
  print(summary(glm_fit))
  print(exp(cbind(OR=coef(glm_fit), suppressMessages(confint(glm_fit)))))
  cat('\nVIF:\n')
  print(car::vif(glm_fit))
  cat('\n')
}
```

### Combining Training Methods and Devices

For sake of completeness, let's also consider a simplified model where the use
of any punishing devices indicates the use of punishment training methods. The
`adj_train_technique` columns exists to represent this.

```{r}
df_exp <- df_exp %>%
  mutate(adj_reward = ifelse(reward == TRUE, TRUE, FALSE))
summary(df_exp$adj_reward)

glm_attribs <- c(
  common_params,
  'adj_reward'
)
```

Now let's build the models.

```{r}
set.seed(1)
for (outcome in outcomes) {
  cat(paste(replicate(80, '-'), collapse=''))
  cat(paste0('\n', outcome, '\n'))
  f <- as.formula(paste0(outcome, '~', '.'))
  
  df_tmp <- df_exp[,c(outcome, glm_attribs)]
  df_tmp <- apply_min_xtab(df_tmp, outcome)
  
  glm_fit <- glm(f, data=df_tmp, family='binomial')
  print(summary(glm_fit))
  print(exp(cbind(OR=coef(glm_fit), suppressMessages(confint(glm_fit)))))
  cat('\nVIF:\n')
  print(car::vif(glm_fit))
  cat('\n')
}
```

# Discussion

In brief, I don't think there is any benefit at looking at the non-punishing
devices. Those devices are staples in the average canine household. We are more
concerned with the effect of the punishing devices. I also believe that the
training method and devices should stay separate to avoid over simplifying the
model and the associated risk of warping the meaning of the collected data.

For these reasons, I believe the second model whihc looks only at the punishing
devices, is the most logical fit for this data set.
